Display class now handles logical -> physical mapping. Added convenience methods to address LEDs by row/col. Also the words are now stored more efficiently. Use (more) explicit data types. Serial speed can now be set in config.

+++ include/animation.h
@@ -7,7 +7,7 @@ class Animation
 {
 public:
     static void SetAnimation(String url);
-    static void SetAnimationFrame(long time);
+    static void SetAnimationFrame(uint32_t time);
 private:
 };
 #endif
